pyproject.toml
[project]
name = "untangling-multi-concern-commits-using-slm"
version = "0.1.0"
description = "Concern is All You Need: Semantic Untangling with Small Language Models"
readme = "README.md"
authors = [
    { name = "gobeumsu", email = "gobeumsu@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    # Core ML/Data Libraries
    "torch",
    "transformers",
    "datasets>=2.15.0",
    "pandas>=2.1.0",
    "numpy",
    # Hugging Face & API Clients
    "huggingface_hub>=0.19.0",
    "openai>=1.0.0",
    "aiohttp>=3.8.0",
    # Configuration & Utilities
    "pyyaml",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "tiktoken>=0.4.0",
    # Visualization
    "matplotlib",
    "seaborn",
    # Streamlit
    "streamlit>=1.28.0",
    "lmstudio",
    "outlines>=1.2.0",
    "accelerate>=0.25.0",
    # GGUF runtime
    "llama-cpp-python>=0.2.79",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["utils", "visual_eval"]
